fix(aggregate): Make SUM() return REAL (float) for SQLite compatibility (#4449) (#4457)
SQLite's SUM() aggregate function always returns a REAL (floating-point)
type, regardless of the input column type. This differs from TOTAL()
which also returns 0.0 for empty sets.

This change updates all SUM() code paths to return SqlValue::Numeric
instead of SqlValue::Integer when aggregating integer values:

- Row-based aggregation (aggregates.rs finalize())
- Columnar aggregation (functions.rs, group_by.rs, mod.rs)
- SIMD aggregation (simd_aggregate.rs, simd.rs)
- Fused execution (fused.rs, aggregate.rs, expression.rs)
- Batch processing (batch.rs, vectorized.rs)

All test assertions have been updated to expect Numeric type for SUM
results, matching SQLite's behavior.

Before: SELECT SUM(x), TYPEOF(SUM(x)) FROM t1; → 6, integer
After:  SELECT SUM(x), TYPEOF(SUM(x)) FROM t1; → 6.0, real

+++ crates/vibesql-executor/src/select/columnar/executor/fused.rs
@@ -164,11 +164,12 @@ fn compute_fused_i64_aggregate(
 ) -> Result<SqlValue, ExecutorError> {
     match op {
         AggregateOp::Sum => {
+            // SQLite's SUM() always returns REAL (float)
             let count = simd_ops::count_filtered(filter_mask);
             if count == 0 {
                 Ok(SqlValue::Null)
             } else {
-                Ok(SqlValue::Integer(simd_ops::sum_i64_filtered(values, filter_mask)))
+                Ok(SqlValue::Numeric(simd_ops::sum_i64_filtered(values, filter_mask) as f64))
             }
         }
         AggregateOp::Count => Ok(SqlValue::Integer(simd_ops::count_filtered(filter_mask))),
